scripts/training_requirements.txt
torch
torchaudio
psycopg2-binary
pandas
numpy
python-dotenv
torch-audiomentations